move config to a separate file
